Recognition of rank-1 matrices.
